backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
pydantic[email]==2.10.2
pydantic-settings==2.6.1
pyjwt==2.9.0
bcrypt==4.2.0
firebase-admin==6.5.0
google-cloud-firestore==2.19.0
